+++ src/vws/exceptions/base_exceptions.py
@@ -5,7 +5,7 @@ Cloud Recognition Web API.
 
 from beartype import beartype
 
-from .response import Response
+from vws.types import Response
 
 
 @beartype